Mensajeria Email e Integridad Operaciones.

+++ Entidades/DetalleOperacion.cs
@@ -9,9 +9,9 @@ namespace Entidades
     {
         public int OperacionId { get; set; }
         public int ProductoId { get; set; }
-        public decimal Monto { get; set; }
+        public int Monto { get; set; }
         public int Cantidad { get; set; }
-        public decimal SubTotal { get; set; }
+        public int SubTotal { get; set; }
         public long DVH { get; set; }
     }
 }
